Added breaks calculation, converted time to 24 hour clock, and fixed tutorial combiner
Each time a section is added to a day in a schedule, a method is called
which calculates the total amount of breaks in the day. The course
meeting times have been converted from a 12 hour format to a 24 hour
format. The tutorial combiner method now checks for the same section and
time (before it only combined tutorials of the same section).