The script now finds a schedule with minimal breaks. The conflict and break calculations were fixed.
A nested for loop is used to check every combination of lectures, labs,
and tutorials. Dummy sections are used if labs or tutorials are missing
in courses. A schedule with no conflicts and minimal breaks is found,
although this process is still quick and dirty. The conflict calculation
thought every schedule had a conflict before, this has been fixed. The
break calculation was broken and would return negative break values
occasionally, this has also been fixed.